refactor(dev): switch to `mod.rs` module root style and reorganize `exec` module
