+++ src/eval/dispatcher.rs
@@ -1,11 +1,10 @@
 use std::os::unix::prelude::PermissionsExt;
 
-use crate::exec::builtins;
-use crate::exec::commands::{Builtin, Executable, Runnable};
-use crate::state::{Path, ShellState};
-
 use super::tokenizer::tokenize;
 use crate::errors::Result;
+use crate::exec::{builtin_funcs, Builtin};
+use crate::exec::{Executable, Runnable};
+use crate::state::{Path, ShellState};
 
 /// Represents a collection of builtin commands
 /// Allows for command resolution and execution through aliases
@@ -19,22 +18,22 @@ impl Default for Dispatcher {
     fn default() -> Self {
         let mut dispatcher = Self::new();
 
-        dispatcher.add_builtin("test", vec!["t"], builtins::test);
-        dispatcher.add_builtin("exit", vec!["quit", "q"], builtins::exit);
-        dispatcher.add_builtin("working-directory", vec!["pwd", "wd"], builtins::working_directory);
-        dispatcher.add_builtin("change-directory", vec!["cd"], builtins::change_directory);
-        dispatcher.add_builtin("list-directory", vec!["directory", "list", "ls", "dir"], builtins::list_directory);
-        dispatcher.add_builtin("previous-directory", vec!["back", "b", "prev", "pd"], builtins::previous_directory);
-        dispatcher.add_builtin("next-directory", vec!["forward", "f", "next", "nd"], builtins::next_directory);
-        dispatcher.add_builtin("clear-terminal", vec!["clear", "cls"], builtins::clear_terminal);
-        dispatcher.add_builtin("make-file", vec!["create", "touch", "new", "mf"], builtins::make_file);
-        dispatcher.add_builtin("make-directory", vec!["mkdir", "md"], builtins::make_directory);
-        dispatcher.add_builtin("delete-file", vec!["delete", "remove", "rm", "del", "df"], builtins::delete_file);
-        dispatcher.add_builtin("read-file", vec!["read", "cat", "rf"], builtins::read_file);
-        dispatcher.add_builtin("run-executable", vec!["run", "exec", "re"], builtins::run_executable);
-        dispatcher.add_builtin("configure", vec!["config", "conf"], builtins::configure);
-        dispatcher.add_builtin("environment-variable", vec!["environment", "env", "ev"], builtins::environment_variable);
-        dispatcher.add_builtin("edit-path", vec!["path", "ep"], builtins::edit_path);
+        dispatcher.add_builtin("test", vec!["t"], builtin_funcs::test);
+        dispatcher.add_builtin("exit", vec!["quit", "q"], builtin_funcs::exit);
+        dispatcher.add_builtin("working-directory", vec!["pwd", "wd"], builtin_funcs::working_directory);
+        dispatcher.add_builtin("change-directory", vec!["cd"], builtin_funcs::change_directory);
+        dispatcher.add_builtin("list-directory", vec!["directory", "list", "ls", "dir"], builtin_funcs::list_directory);
+        dispatcher.add_builtin("previous-directory", vec!["back", "b", "prev", "pd"], builtin_funcs::previous_directory);
+        dispatcher.add_builtin("next-directory", vec!["forward", "f", "next", "nd"], builtin_funcs::next_directory);
+        dispatcher.add_builtin("clear-terminal", vec!["clear", "cls"], builtin_funcs::clear_terminal);
+        dispatcher.add_builtin("make-file", vec!["create", "touch", "new", "mf"], builtin_funcs::make_file);
+        dispatcher.add_builtin("make-directory", vec!["mkdir", "md"], builtin_funcs::make_directory);
+        dispatcher.add_builtin("delete-file", vec!["delete", "remove", "rm", "del", "df"], builtin_funcs::delete_file);
+        dispatcher.add_builtin("read-file", vec!["read", "cat", "rf"], builtin_funcs::read_file);
+        dispatcher.add_builtin("run-executable", vec!["run", "exec", "re"], builtin_funcs::run_executable);
+        dispatcher.add_builtin("configure", vec!["config", "conf"], builtin_funcs::configure);
+        dispatcher.add_builtin("environment-variable", vec!["environment", "env", "ev"], builtin_funcs::environment_variable);
+        dispatcher.add_builtin("edit-path", vec!["path", "ep"], builtin_funcs::edit_path);
 
         dispatcher
     }
