--- src/exec.rs
@@ -1,3 +1,0 @@
-mod builtin_arguments;
-pub mod builtins;
-pub mod commands;
