+++ src/exec/mod.rs
@@ -1,0 +1,8 @@
+mod builtins;
+mod executable;
+mod runnable;
+
+pub use builtins::command::Builtin;
+pub use builtins::functions as builtin_funcs;
+pub use executable::Executable;
+pub use runnable::Runnable;
